implement update method to user update their password and some also add index method to see all user

+++ app/controllers/users_controller.rb
@@ -2,9 +2,15 @@ class UsersController < ApplicationController
   # protect_from_forgery with: :null_session
   skip_before_action :verify_authenticity_token
 
+  def index
+    render json: User.all
+  end
+
   def login
     @user = User.find_by(email: params[:email], password: params[:password])
-    if @user
+    if session[:current_user_id] != nil
+      render json: { message: "Already login" }
+    elsif @user
       session[:current_user_id] = @user.id
       render json: { message: "Successfully Login"}
     else
@@ -21,14 +27,18 @@ class UsersController < ApplicationController
       if @user.save
         render json: @user
       else
-        render :new, status: "error"
+        render json: { message: "Invalid" }
       end
     end
   end
 
   def logout
-    session.delete(:current_user_id)
-    render json: { message: "Logout"}
+    if session[:current_user_id] != nil
+      session.delete(:current_user_id)
+      render json: { message: "Logout"}
+    else
+      render json: { message: "Login required" }
+    end
   end
 
   def profile
@@ -39,8 +49,22 @@ class UsersController < ApplicationController
     end
   end
 
+  def update
+    if session[:current_user_id] != nil
+      @user = User.find_by_id(session[:current_user_id])
+
+      if @user.update_attribute(:password, params[:password]) && params[:password].present?
+        render json: User.find_by_id(session[:current_user_id])
+      else
+        render json: { message: "Something went wrong" }
+      end
+    else
+      render json: { message: "Login required" }
+    end
+  end
+
   private
     def user_param
-      params.require(:user).permit(:email, :password)
+      params.permit(:email, :password)
     end
 end
